+++ packages/terra-application/src/application-base/private/getBrowserLocale.js
@@ -49,6 +49,10 @@ const getBrowserLocale = () => {
   if (preferredLocale) {
     return preferredLocale;
   }
+  /* for IE support, as languages and language in IE return undefined, and userLanguage and browserLanguage return "en-US" */
+  if (isSupported(navigator.systemLanguage)) {
+    return navigator.systemLanguage;
+  }
 
   if (isSupported(navigator.language)) {
     return navigator.language;